@@ -32,68 +32,74 @@
     "test": "ng test --code-coverage",
     "ngE2e": "sleepms 1000 && ng e2e",
     "e2e": "npm-run-all -p -r startServer ngE2e",
-    "resetDatabase": "ts-node --project ./utils/tsconfig.utils.json ./utils/reset-database.ts run"
+    "resetTestDatabase": "ts-node --project ./utils/tsconfig.utils.json ./e2e/server-utils/reset-testDatabase.ts run"
   },
   "dependencies": {
-    "@angular/animations": "^8.1.1",
-    "@angular/cdk": "~8.0.2",
-    "@angular/common": "^8.1.1",
-    "@angular/compiler": "^8.1.1",
-    "@angular/core": "^8.1.1",
+    "@angular/animations": "^8.2.5",
+    "@angular/cdk": "^8.1.4",
+    "@angular/common": "^8.2.5",
+    "@angular/compiler": "^8.2.5",
+    "@angular/core": "^8.2.5",
     "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "^8.1.1",
-    "@angular/material": "^8.0.2",
-    "@angular/platform-browser": "^8.1.1",
-    "@angular/platform-browser-dynamic": "^8.1.1",
-    "@angular/router": "^8.1.1",
-    "@angular/upgrade": "^8.1.1",
-    "@types/protractor-browser-logs": "^1.0.9",
-    "@types/request-promise-native": "^1.0.16",
-    "core-js": "^3.1.4",
+    "@angular/forms": "^8.2.5",
+    "@angular/material": "^8.1.4",
+    "@angular/platform-browser": "^8.2.5",
+    "@angular/platform-browser-dynamic": "^8.2.5",
+    "@angular/router": "^8.2.5",
+    "@angular/upgrade": "^8.2.5",
+    "@auth0/auth0-spa-js": "^1.2.3",
+    "core-js": "^3.2.1",
     "hammerjs": "^2.0.8",
     "http-status-codes": "^1.3.2",
-    "ngx-logger": "^3.4.5",
-    "ngx-toastr": "^10.0.4",
-    "rollbar": "^2.8.1",
-    "rxjs": "^6.5.2",
-    "zone.js": "~0.9.1"
+    "ngx-logger": "^4.0.4",
+    "ngx-toastr": "^10.2.0",
+    "rollbar": "^2.12.3",
+    "rxjs": "^6.5.3",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.801.1",
-    "@angular/cli": "^8.1.1",
-    "@angular/compiler-cli": "^8.1.1",
-    "@angular/platform-server": "^8.1.1",
+    "@angular-devkit/build-angular": "~0.803.3",
+    "@angular/cli": "^8.3.3",
+    "@angular/compiler-cli": "^8.2.5",
+    "@angular/platform-server": "^8.2.5",
     "@types/app-root-path": "^1.2.4",
-    "@types/jasmine": "^3.3.13",
+    "@types/dotenv": "^6.1.1",
+    "@types/jasmine": "^3.4.0",
     "@types/jasminewd2": "^2.0.6",
-    "@types/node": "^12.6.2",
+    "@types/node": "^12.7.4",
+    "@types/protractor-browser-logs": "^1.0.9",
+    "@types/request-promise-native": "^1.0.16",
     "@types/shelljs": "^0.8.5",
     "app-root-path": "^2.2.1",
     "codelyzer": "^5.1.0",
+    "dotenv": "^8.1.0",
     "jasmine-core": "~3.4.0",
+    "jasmine-reporters": "^2.3.2",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.5",
+    "karma": "~4.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
     "karma-phantomjs-launcher": "^1.0.4",
-    "lodash": "^4.17.14",
+    "lodash": "^4.17.15",
     "node-sass": "^4.12.0",
     "npm-run-all": "^4.1.5",
     "phantomjs-prebuilt": "^2.1.16",
     "prettier": "^1.18.2",
     "protractor": "^5.4.2",
     "protractor-browser-logs": "^1.0.456",
+    "protractor-html-reporter": "^1.3.2",
     "request": "^2.88.0",
     "request-promise-native": "^1.0.7",
     "shelljs": "^0.8.3",
     "sleep-ms": "^2.0.1",
     "ts-node": "~8.3.0",
-    "tslint": "~5.18.0",
+    "tslint": "~5.19.0",
     "tslint-angular": "^3.0.2",
     "tslint-no-unused-expression-chai": "^0.1.4",
-    "typescript": "^3.4.5",
-    "wait-on": "^3.3.0"
+    "typescript": "^3.5.1",
+    "wait-on": "^3.3.0",
+    "webdriver-manager": "^12.1.6"
   }
 }
