@@ -37,6 +37,7 @@ function notFound(_req: IRequestApp, _res: Response, next: NextFunction) {
 }
 
 /**
+ * Catches all errors.
  * Asssigns a HTTP error code to res.statusCode.
  * A code is assigned to communicate a message to the client.
  * Note: Errors left with code 500 (Internal Server Error) trigger a uncaught
@@ -51,14 +52,24 @@ function assignCode(
 ) {
   debug(modulename + ': assignCode called');
 
-  if (typeof err === 'object') {
-    /* set the response status code to the error statusCode field, if one was added on error creation, or if one was added above. */
-    res.statusCode = err.statusCode ? err.statusCode : res.statusCode;
-    /* If empty, set the response code to internal server error, 500. */
-    res.statusCode = res.statusCode ? res.statusCode : 500;
-    /* If 20x, set the response code to internal server error, 500. */
-    res.statusCode =
-      res.statusCode < 200 || res.statusCode > 299 ? res.statusCode : 500;
+  if (typeof err !== 'object') {
+    err = new Error(String(err));
+  }
+
+  /* set the response status code to the error statusCode field, if one was added on error creation, or if one was added above. */
+  res.statusCode = err.statusCode ? err.statusCode : res.statusCode;
+  /* if empty, set the response code to internal server error, 500. */
+  res.statusCode = res.statusCode ? res.statusCode : 500;
+  /* if 20x, set the response code to internal server error, 500. */
+  res.statusCode =
+    res.statusCode < 200 || res.statusCode > 299 ? res.statusCode : 500;
+
+  /* set mandatory header for 401 */
+  if (res.statusCode === 401) {
+    res.setHeader(
+      'WWW-Authenticate',
+      'OAuth realm="Access to team members API", charset="UTF-8"',
+    );
   }
 
   next(err);
