@@ -6,7 +6,7 @@
  * This module calls client-side tests that test server error handling.
  */
 
-/* tests server errorhandler */
+/* tests server error handler */
 import '/testServer/errors/errors.client-test.js';
 
 mocha.checkLeaks();
