@@ -1,23 +1,9 @@
 import { browser } from 'protractor';
 
-import { getRootElement } from './elements/root.element';
-import { getDashboardElement } from './elements/dashboard.element';
+import { getRootElements } from './elements/root.elements';
+import { getDashboardElements } from './elements/dashboard.elements';
 import { getMemberSearchElement } from './elements/memberSearch.element';
-import { getMessagesElement } from './elements/messages.element';
-import { getToastrElement } from './elements/toastr.element';
-
-/**
- * A get page function exists for each page:
- * - dashboard page containing app-root, app-dashboard, app-member-search and app-messages.
- * - members page containing app-root, app-members and app-messages.
- * - member detail page containing app-root, app-member-detail and app-messages.
- * - page not found page containing app-root and app-page-not-found.
- *
- * The elements that appear on the page are exported, along with helper functions.
- *
- * The elements need to be on the DOM when an action is called, i.e. only call the get dashboard page function when the dashboard page is being displayed.
- *
- */
+import { getToastrElements } from './elements/toastr.elements';
 
 export function getDashboardPage() {
 
@@ -26,18 +12,16 @@ export function getDashboardPage() {
     await browser.get('/dashboard');
   }
 
-  const rootElement = getRootElement();
-  const dashboardElement = getDashboardElement();
+  const rootElements = getRootElements();
+  const dashboardElements = getDashboardElements();
   const memberSearchElement = getMemberSearchElement();
-  const messagesElement = getMessagesElement();
-  const toastrElement = getToastrElement();
+  const toastrElements = getToastrElements();
 
   return {
     navigateToPage,
-    rootElement,
-    dashboardElement,
+    rootElements,
+    dashboardElements,
     memberSearchElement,
-    messagesElement,
-    toastrElement,
+    toastrElements,
   }
 }
